refactor(cli): extract db inspect+profiling commands to _db_profiling
B8 of the Round 2 plan. Moves cmd_inspect, cmd_profiling, and the
_format_age helper out of amx/cli_support/commands/db.py into a
focused amx/cli_support/_db_profiling.py.

cmd_inspect diagnoses an active or named DB profile (backend +
connection + capability flags + visible schemas/table counts); the
147-LOC handler is the longest in db.py. cmd_profiling shows or
updates the per-profile profiling guardrails. _format_age formats
cache freshness for both.

db.py re-exports each name. test_profiling_cli_approximate's patch
of amx.cli_support.commands.db.error rerouted to the new module.

db.py: 1807 -> 1572 LOC (-235, -13%).
